+++ src/serviceDefinition/apply.ts
@@ -1,0 +1,7 @@
+import { uploadServiceDefinition } from './upload';
+import { loadServiceDefinition } from './loading';
+
+export async function apply(serviceDefinitionPath: string) {
+  const serviceDefinition = await loadServiceDefinition(serviceDefinitionPath);
+  await uploadServiceDefinition(serviceDefinition);
+}