+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 ## Unreleased
+ - Rename `contract` to `service definition`
+ - Support a `delete` command.
+ - Rename `establish` to `apply` to be closer to the wording of kubectl.
+ - Add `steadybit service` as an alias for `steadybit service-definition`.
+
+## v0.2.0
  - Add validation to the `contract init` command.
  - Adapt generated yaml structure to new format.
  - Improve documentation.